chore: Address claude-review feedback on PR 1243
- fix(core): _format_bound now keeps full precision for sub-1e-6 bounds.
  format(value, "f") defaults to 6 fractional digits, so format(1e-7,"f")
  rstripped to "0", silently widening declared bounds. Use repr() for the
  shortest exact-roundtrip representation, falling back to .17f when repr
  switches to scientific notation. Adds a regression test at min=1e-7.

- doc(arc42): ADR documenting the Nuxt 3.21.0 hard pin (vite-node IPC
  regression) with an explicit exit criterion so the pin doesn't become
  forgotten tech debt.

- refactor(ui): Extract identical Agent/Configuration.vue and
  Process/Configuration.vue into FormKit/DynamicConfiguration.vue. Both
  components become thin wrappers; future form-lifecycle changes only need
  to be made once.

- doc(ui): Comment seedFormDefaults on its testability gap (vitest not yet
  configured) and on the value === undefined group branch (relies on
  coerceNullableToggles to nullify at submit). Document the
  unique-by-construction assumption behind the cleanNode.key fallback.
