+++ packages/core/swiss_ai_hub/core/form/elements/input_number.py
@@ -59,11 +59,19 @@ class InputNumber(PrimeVueElement):
 
     @staticmethod
     def _format_bound(value: float) -> str:
-        # Render integer-valued floats as "5" not "5.0", but avoid `:g`'s scientific
-        # notation for large magnitudes — FormKit's min/max rules can't parse "1e+06".
+        # FormKit's min/max rules can't parse scientific notation like "1e-07" or "1e+06",
+        # so emit a plain decimal in every case.
         if value.is_integer():
             return str(int(value))
-        return format(value, "f").rstrip("0").rstrip(".")
+        # `repr(float)` returns the shortest exact-roundtrip representation (e.g.
+        # repr(1234567.89) == '1234567.89' with no precision artefacts), but switches
+        # to scientific notation for magnitudes below ~1e-4. When that happens, fall
+        # back to a fixed-point format with enough precision to roundtrip any IEEE-754
+        # double (17 significant digits).
+        formatted = repr(value)
+        if "e" in formatted or "E" in formatted:
+            formatted = format(value, ".17f")
+        return formatted.rstrip("0").rstrip(".")
 
     def in_locale(self, t: LocaleHandler) -> Self:
         self_copy = super().in_locale(t)
